update forge fmt

+++ test/interface.sol
@@ -12,40 +12,40 @@ import {IERC20} from "forge-std/interfaces/IERC20.sol";
  * IWETH
  * IUNI_ROUTER_V3
  * IBalancer
-*/
+ */
 
 interface IMorpho {
     function flashLoan(address token, uint256 assets, bytes calldata data) external;
 }
 
 interface IDodo {
-    function flashLoan (uint256 baseAmount, uint256 quoteAmount, address assetTo, bytes calldata data) external;
+    function flashLoan(uint256 baseAmount, uint256 quoteAmount, address assetTo, bytes calldata data) external;
 }
 
 interface IPancakeV2Router {
     function swapExactTokensForTokensSupportingFeeOnTransferTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external;
     function getAmountsOut(uint256 amountIn, address[] calldata path) external view returns (uint256[] memory amounts);
 }
-
 
 interface IPancakeLP {
     function getReserves() external view returns (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast);
 }
 
-interface IAAVE_V3{
-    function flashLoanSimple(address receiver, address token, uint256 amount, bytes memory params, uint16 referralCode) external;
+interface IAAVE_V3 {
+    function flashLoanSimple(address receiver, address token, uint256 amount, bytes memory params, uint16 referralCode)
+        external;
 }
 
-interface IUNI_ROUTER_V3{
-    struct Params{
+interface IUNI_ROUTER_V3 {
+    struct Params {
         address tokenIn;
-        address tokenOut; 
+        address tokenOut;
         uint24 fee;
         address recipient;
         uint256 deadline;
@@ -53,19 +53,18 @@ interface IUNI_ROUTER_V3{
         uint256 amountOutMinimum;
         uint160 sqrtPriceLimitX96;
     }
-    
+
     function exactInputSingle(Params calldata params) external returns (uint256);
 }
 
-interface IWETH{
+interface IWETH {
     function withdraw(uint256 amount) external;
-    function balanceOf(address account) external returns(uint256);
-    function decimals() external returns(uint256);
+    function balanceOf(address account) external returns (uint256);
+    function decimals() external returns (uint256);
     function deposit() external payable;
-    function transfer(address dst, uint wad) external returns (bool);
-    function approve(address guy, uint wad) external returns (bool);
+    function transfer(address dst, uint256 wad) external returns (bool);
+    function approve(address guy, uint256 wad) external returns (bool);
 }
-
 
 interface IFlashLoanRecipient {
     function receiveFlashLoan(
@@ -75,7 +74,8 @@ interface IFlashLoanRecipient {
         bytes memory userData
     ) external;
 }
-interface IBalancerVault{
+
+interface IBalancerVault {
     function flashLoan(
         IFlashLoanRecipient recipient,
         IERC20[] memory tokens,
